% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsaf_aedplot.R
\name{labsaf_eadplot}
\alias{labsaf_eadplot}
\title{labsaf aed}
\usage{
labsaf_eadplot(data, var, fill, mode = "plot")
}
\arguments{
\item{data, }{is an sf or data.frame object with the values of the \strong{soil properties}.}

\item{var, }{represent the values of the soil properties.}

\item{fill, }{is a fill color hexadecimal for plot.}

\item{mode, }{is the output mode static o view plot.}
}
\value{
A graphic of ggplot2.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a function to create a some plot of analysis exploratory of data.
}
