% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsaf_geoplot.R
\name{labsaf_geoplot}
\alias{labsaf_geoplot}
\title{labsaf geoplot}
\usage{
labsaf_geoplot(region, data, var, ...)
}
\arguments{
\item{region, }{is an sf object of type polygon that represent the study area.}

\item{data, }{is an sf object with the values of the \strong{soil properties}.}

\item{var, }{represent the values of the soil properties.}

\item{..., }{further arguments of \code{labs} of the ggplot2 package.}
}
\value{
A graphic of ggplot2.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a function to create a point map with the values of the soil properties.
}
