% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsaf_make_grid.R
\name{labsaf_make_grid}
\alias{labsaf_make_grid}
\title{labsaf make grid}
\usage{
labsaf_make_grid(x, size)
}
\arguments{
\item{x, }{is a sf object (polygon) that represent the study area.}

\item{size, }{is the size of the pixel grid.}
}
\value{
a sf object (polygons).
}
\description{
Is a function to create a grid of study area.
}
