% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsaf_make_grid.R
\name{labsaf_make_grid}
\alias{labsaf_make_grid}
\title{labsaf make grid}
\usage{
labsaf_make_grid(x, size, square = TRUE, ...)
}
\arguments{
\item{x, }{is as sf object of type polygon (\emph{UTM project}) that represent the study area.}

\item{size, }{is the size of the pixel grid in \strong{meters}.}

\item{square, }{is logical; if \code{FALSE}, create \strong{hexagonal grid}.}

\item{..., }{further arguments of \code{st_make_grid()} of the sf package.}
}
\value{
An sf object of type polygons.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a function to create a grid regular o hexagonal of the study area.
}
